@@ -1,1 +1,1 @@
-self.__BUILD_MANIFEST={polyfillFiles:["static/chunks/polyfills-78c92fac7aa8fdd8.js"],devFiles:[],ampDevFiles:[],lowPriorityFiles:["static/HnejM3rzFgOrFHLXVpGh6/_buildManifest.js","static/HnejM3rzFgOrFHLXVpGh6/_ssgManifest.js"],rootMainFiles:["static/chunks/webpack-ce8f8b7286a7c285.js","static/chunks/fd9d1056-a2d8415e863ee978.js","static/chunks/23-a08ee6cf9ea1b421.js","static/chunks/main-app-00191d8baf8ff448.js"],pages:{"/_app":["static/chunks/webpack-ce8f8b7286a7c285.js","static/chunks/framework-841a78cfa510b3f8.js","static/chunks/main-55e53cdb98ca49b5.js","static/css/17b39a50e277f2a5.css","static/chunks/pages/_app-e85989cb719f9630.js"],"/_error":["static/chunks/webpack-ce8f8b7286a7c285.js","static/chunks/framework-841a78cfa510b3f8.js","static/chunks/main-55e53cdb98ca49b5.js","static/chunks/pages/_error-1be831200e60c5c0.js"]},ampFirstPages:[]};
+self.__BUILD_MANIFEST={polyfillFiles:["static/chunks/polyfills-78c92fac7aa8fdd8.js"],devFiles:[],ampDevFiles:[],lowPriorityFiles:["static/9sCa8WB3aIuoZOshcT-hl/_buildManifest.js","static/9sCa8WB3aIuoZOshcT-hl/_ssgManifest.js"],rootMainFiles:["static/chunks/webpack-ce8f8b7286a7c285.js","static/chunks/fd9d1056-a2d8415e863ee978.js","static/chunks/23-a08ee6cf9ea1b421.js","static/chunks/main-app-00191d8baf8ff448.js"],pages:{"/_app":["static/chunks/webpack-ce8f8b7286a7c285.js","static/chunks/framework-841a78cfa510b3f8.js","static/chunks/main-55e53cdb98ca49b5.js","static/css/8807443c1787cf20.css","static/chunks/pages/_app-7378c5693af5e836.js"],"/_error":["static/chunks/webpack-ce8f8b7286a7c285.js","static/chunks/framework-841a78cfa510b3f8.js","static/chunks/main-55e53cdb98ca49b5.js","static/chunks/pages/_error-1be831200e60c5c0.js"]},ampFirstPages:[]};
