@@ -1,5 +1,6 @@
-// tailwind.config.js
-module.exports = {
+import type { Config } from 'tailwindcss';
+
+const config: Config = {
   content: [
     './pages/**/*.{js,ts,jsx,tsx}',
     './components/**/*.{js,ts,jsx,tsx}',
@@ -56,3 +57,5 @@ module.exports = {
   },
   plugins: [],
 };
+
+export default config;
